refactor: Improve registerUser test for duplicate usernames
This commit refactors the 'should not allow duplicate username' test
in AuthService.test.ts. The assignment of the result of
`AuthService.registerUser` to `anotherUser` was removed, as the variable
wasn't used. The test now focuses solely on asserting that the
`registerUser` function throws an AppError when attempting to register
a user with an existing username, improving the test's clarity and
reducing unnecessary code.

Signed-off-by: kallyas <kallyasmedia@gmail.com>