feat: Add Pods project to Xcode workspace
This commit adds the Pods project (Pods/Pods.xcodeproj) to the Xcode workspace configuration file (contents.xcworkspacedata). This ensures that the Pods project is properly integrated into the workspace, allowing Xcode to build and link against the dependencies managed by CocoaPods. This is a standard step when integrating CocoaPods into an iOS project.

Signed-off-by: kallyas <kallyasmedia@gmail.com>