feat: Add JWT token to socket.io authentication
This commit configures the socket.io client to include the JWT token
in the authentication headers.  Specifically, it adds `.setAuth({'token': token})`
to the `IO.OptionsBuilder` to send the token as part of the socket connection
handshake. This allows the server to authenticate the client based on the token.
This supplements the existing `authorization` header.

Signed-off-by: kallyas <kallyasmedia@gmail.com>