fix: Include Generated.xcconfig in Debug build configuration
This commit ensures that the `Generated.xcconfig` file is included
in the Debug build configuration for the iOS Runner project. This
inclusion is necessary for proper build setup and to resolve
potential issues related to missing build settings or configurations
defined within `Generated.xcconfig`. Resolves build issues related
to missing configurations in debug builds.

Signed-off-by: kallyas <kallyasmedia@gmail.com>