fix: Update dependencies in yarn.lock to resolve conflicts
This commit updates the `yarn.lock` file to resolve dependency
conflicts and inconsistencies. Specifically:

- Removed redundant version specifiers for packages like `@babel/core`,
  `@jest/transform`, `@jest/types`, `eslint`, `babel-jest`, and
  `typescript` to ensure a single, consistent version is used.
- Reordered version specifiers for `@colors/colors`,
 `base64id`,`@nodelib/fs.stat`,`@types/node`,`qs`,`safe-buffer`,`source-map-support`,`unpipe` packages.
- Reordered and added "@jridgewell/trace-mapping" versions.
- Removed duplicated `camelcase` entry, and `string_decoder`,`find-up`
`minimatch`, `ms` and `glob-parent`.

These changes ensure that the project uses consistent and compatible
dependency versions, reducing the risk of unexpected behavior or bugs.
The update of "jest" and "babel" related dependencies are targeted for
testing framework.

Signed-off-by: kallyas <kallyasmedia@gmail.com>